moved adapters to external folder
