+++ adapters/Behat2/ListFeaturesExtension/Console/Processor/ListFeaturesProcessor.php
@@ -1,13 +1,16 @@
 <?php
 
-namespace Liuggio\Fastest\Behat\ListFeaturesExtension\Console\Processor;
+namespace Liuggio\Fastest\Behat2\ListFeaturesExtension\Console\Processor;
 
 use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Behat\Behat\Console\Processor\Processor as BaseProcessor;
+use Symfony\Component\DependencyInjection\ContainerInterface,
+    Symfony\Component\Console\Command\Command,
+    Symfony\Component\Console\Input\InputArgument,
+    Symfony\Component\Console\Input\InputInterface,
+    Symfony\Component\Console\Output\OutputInterface;
+
+
+use  Behat\Behat\Console\Processor\Processor as BaseProcessor;
 
 /**
  * Extends Behat command line options to provide options that lists the individual feature files / scenarios that
@@ -52,6 +55,7 @@ class ListFeaturesProcessor extends BaseProcessor
      */
     public function process(InputInterface $input, OutputInterface $output)
     {
+
         $listFeatures = $input->getOption('list-features');
         $listScenarios = $input->getOption('list-scenarios');
 
@@ -59,12 +63,17 @@ class ListFeaturesProcessor extends BaseProcessor
             return;
         }
 
-        $featurePath = $input->getArgument('features') ?: $this->container->getParameter('behat.paths.features');
         $results = [];
-        if ($listFeatures) {
-            $results = array_merge($results, $this->_getFeatureFiles($featurePath));
-        } else {
-            $results = array_merge($results, $this->_getScenarios($featurePath));
+        $featurePaths = $this->container->get('behat.console.command')->getFeaturesPaths();
+        foreach($featurePaths as $featurePath) {
+
+            if ($listFeatures) {
+                $results = array_merge($results, $this->_getFeatureFiles($featurePath));
+            }
+            else {
+                $results = array_merge($results, $this->_getScenarios($featurePath));
+            }
+
         }
 
         $output->writeln(implode(PHP_EOL, $results));
@@ -85,13 +94,15 @@ class ListFeaturesProcessor extends BaseProcessor
         $scenarios = [];
         /* @var $featureNodes \Behat\Gherkin\Node\FeatureNode[] */
         $featureNodes = $this->container->get('gherkin')->load($featurePath);
-        foreach ($featureNodes as $featureNode) {
+        foreach($featureNodes as $featureNode) {
+
             /* @var $scenario \Behat\Gherkin\Node\ScenarioNode */
             foreach ($featureNode->getScenarios() as $scenario) {
                 $file = $scenario->getFile();
                 $lineNo = $scenario->getLine();
                 $scenarios[] = "$file:$lineNo";
             }
+
         }
 
         return $scenarios;
@@ -109,7 +120,7 @@ class ListFeaturesProcessor extends BaseProcessor
         $featureFiles = [];
         /* @var $featureNodes \Behat\Gherkin\Node\FeatureNode[] */
         $featureNodes = $this->container->get('gherkin')->load($featurePath);
-        foreach ($featureNodes as $featureNode) {
+        foreach($featureNodes as $featureNode) {
             $featureFiles[] = $featureNode->getFile();
         }
