+++ adapters/Behat2/ListFeaturesExtension/services/core.xml
@@ -3,16 +3,12 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
-
-        <parameter key="liuggio.fastest.behat.list_features_extension.console.processor.list_features.class">Liuggio\Fastest\Behat\ListFeaturesExtension\Console\Processor\ListFeaturesProcessor</parameter>
-
+        <parameter key="liuggio.fastest.behat2.list_features_extension.console.processor.list_features.class">Liuggio\Fastest\Behat2\ListFeaturesExtension\Console\Processor\ListFeaturesProcessor</parameter>
     </parameters>
     <services>
-
-        <service id="liuggio.fastest.behat.list_features.console.processor.list_features" class="%liuggio.fastest.behat.list_features_extension.console.processor.list_features.class%">
+        <service id="liuggio.fastest.behat2.list_features.console.processor.list_features" class="%liuggio.fastest.behat2.list_features_extension.console.processor.list_features.class%">
             <argument type="service" id="service_container" />
             <tag name="behat.console.processor" />
         </service>
-
     </services>
 </container>
