+++ adapters/Doctrine/MongoDB/Connection.php
