+++ adapters/Environment/FastestEnvironment.php
